+++ x-pack/plugins/cloud_security_posture/public/pages/rules/rules_context.tsx
@@ -8,74 +8,79 @@ import React, { createContext, useContext, useMemo, useEffect, useCallback, useS
 import { useParams } from 'react-router-dom';
 
 import {
-  FindCspBenchmarkRuleResponse,
+  CspBenchmarkRule,
   PageUrlParams,
+  RuleStateAttributes,
 } from '@kbn/cloud-security-posture-common/schema/rules/latest';
+import { extractErrorMessage } from '@kbn/cloud-security-posture-common';
+import { buildRuleKey } from '../../../common/utils/rules_states';
 import { RulesQuery, useFindCspBenchmarkRule } from './use_csp_benchmark_rules';
 import { usePageSize } from '../../common/hooks/use_page_size';
 import { LOCAL_STORAGE_PAGE_SIZE_RULES_KEY } from '../../common/constants';
+import { useCspGetRulesStates } from './use_csp_rules_state';
 
-// type Dispatch = (action: Action) => void;
+export interface CspBenchmarkRulesWithStates {
+  metadata: CspBenchmarkRule['metadata'];
+  state: 'muted' | 'unmuted';
+}
+
+interface RulesPageData {
+  rules_page: CspBenchmarkRulesWithStates[];
+  all_rules: CspBenchmarkRulesWithStates[];
+  rules_map: Map<string, CspBenchmarkRulesWithStates>;
+  total: number;
+  error?: string;
+  loading: boolean;
+}
+
+export type RulesState = RulesPageData & RulesQuery;
+
 interface RulesProviderProps {
   children: React.ReactNode;
 }
 interface RulesContextValue {
-  allRules: FindCspBenchmarkRuleResponse | undefined;
   rulesQuery: RulesQuery;
+  rulesPageData: RulesPageData;
   setRulesQuery: (query: Partial<RulesQuery>) => void;
   pageSize: number;
   setPageSize: (pageSize: number) => void;
   page: number;
   setPage: (page: number) => void;
+  sectionList: string[] | undefined;
+  ruleNumberSelectOptions: string[];
+  sectionSelectOptions: string[];
+  selectedRules: CspBenchmarkRulesWithStates[];
+  setSelectedRules: (rules: CspBenchmarkRulesWithStates[]) => void;
+  setSelectAllRules: () => void;
+  setEnabledDisabledItemsFilter: (filter: string) => void;
+  enabledDisabledItemsFilter: string;
+  mutedRulesCount?: number;
+  rulesFlyoutData: CspBenchmarkRulesWithStates;
 }
 
 const RulesContext = createContext<RulesContextValue | undefined>(undefined);
 const MAX_ITEMS_PER_PAGE = 10000;
 
-// interface State {
-//   rulesQuery: RulesQuery;
-//   pageSize?: number;
-//   page: number;
-// }
+const getPage = (data: CspBenchmarkRulesWithStates[], { page, perPage }: RulesQuery) =>
+  data.slice(page * perPage, (page + 1) * perPage);
 
-// type Action =
-//   | {
-//       type: 'setRulesQuery';
-//       value: Partial<RulesQuery>;
-//     }
-//   | { type: 'setPageSize'; value: number }
-//   | { type: 'setPage'; value: number };
+const getRulesPageData = (
+  data: CspBenchmarkRulesWithStates[],
+  status: string,
+  error: unknown,
+  query: RulesQuery
+): RulesPageData => {
+  const page = getPage(data, query);
 
-// function rulesReducer(state: State, action: Action) {
-//   switch (action.type) {
-//     case 'setRulesQuery': {
-//       return { ...state, rulesQuery: { ...state.rulesQuery, ...action.value } };
-//     }
-//     case 'setPage': {
-//       return { ...state, page: action.value };
-//     }
-//     case 'setPageSize': {
-//       return { ...state, pageSize: action.value };
-//     }
-//     default: {
-//       throw new Error(`Unhandled action type: ${action}`);
-//     }
-//   }
-// }
-
-// const initialState: State = {
-//   page: 1,
-//   pageSize: 10,
-//   rulesQuery: {
-//     section: undefined,
-//     ruleNumber: undefined,
-//     search: '',
-//     page: 0,
-//     perPage: 10,
-//     sortField: 'metadata.benchmark.rule_number',
-//     sortOrder: 'asc',
-//   },
-// };
+  return {
+    loading: status === 'loading',
+    error: error ? extractErrorMessage(error) : undefined,
+    all_rules: data,
+    rules_map: new Map(data.map((rule) => [rule.metadata.id, rule])),
+    rules_page: page,
+    total: data?.length || 0,
+  };
+};
 
 export function RulesProvider({ children }: RulesProviderProps) {
   const params = useParams<PageUrlParams>();
@@ -90,6 +95,8 @@ export function RulesProvider({ children }: RulesProviderProps) {
     sortField: 'metadata.benchmark.rule_number',
     sortOrder: 'asc',
   });
+  const [selectedRules, setSelectedRules] = useState<CspBenchmarkRulesWithStates[]>([]);
+  const [enabledDisabledItemsFilter, setEnabledDisabledItemsFilter] = useState('no-filter');
 
   const allRules = useFindCspBenchmarkRule(
     {
@@ -150,28 +157,137 @@ export function RulesProvider({ children }: RulesProviderProps) {
     [setPage]
   );
 
-  const initialContextValue = useMemo(
+  const sectionList = useMemo(
+    () => allRules?.data?.items.map((rule) => rule.metadata.section),
+    [allRules]
+  );
+
+  const ruleNumberSelectOptions = useMemo(
+    () =>
+      allRules.data
+        ? allRules.data.items.map((rule) => rule.metadata.benchmark.rule_number || '')
+        : [],
+    [allRules]
+  );
+
+  const sectionSelectOptions = [...new Set(sectionList)].sort((a, b) => {
+    return a.localeCompare(b, 'en', { sensitivity: 'base' });
+  });
+
+  const { data, status, error } = useFindCspBenchmarkRule(
+    {
+      section: rulesQuery.section,
+      ruleNumber: rulesQuery.ruleNumber,
+      search: rulesQuery.search,
+      page: 1,
+      perPage: MAX_ITEMS_PER_PAGE,
+      sortField: 'metadata.benchmark.rule_number',
+      sortOrder: rulesQuery.sortOrder,
+    },
+    params.benchmarkId,
+    params.benchmarkVersion
+  );
+
+  const rulesStates = useCspGetRulesStates();
+
+  const rulesWithStates: CspBenchmarkRulesWithStates[] = useMemo(() => {
+    if (!data) return [];
+
+    return data.items
+      .filter((rule: CspBenchmarkRule) => rule.metadata.benchmark.rule_number !== undefined)
+      .map((rule: CspBenchmarkRule) => {
+        const rulesKey = buildRuleKey(
+          rule.metadata.benchmark.id,
+          rule.metadata.benchmark.version,
+          /* Rule number always exists* from 8.7 */
+          rule.metadata.benchmark.rule_number!
+        );
+
+        const match = rulesStates?.data?.[rulesKey];
+        const rulesState = match?.muted ? 'muted' : 'unmuted';
+
+        return { ...rule, state: rulesState || 'unmuted' };
+      });
+  }, [data, rulesStates?.data]);
+
+  const mutedRulesCount = rulesWithStates.filter((rule) => rule.state === 'muted').length;
+
+  const filteredRulesWithStates: CspBenchmarkRulesWithStates[] = useMemo(() => {
+    if (enabledDisabledItemsFilter === 'disabled')
+      return rulesWithStates?.filter((rule) => rule?.state === 'muted');
+    else if (enabledDisabledItemsFilter === 'enabled')
+      return rulesWithStates?.filter((rule) => rule?.state === 'unmuted');
+    else return rulesWithStates;
+  }, [rulesWithStates, enabledDisabledItemsFilter]);
+
+  const rulesPageData = useMemo(
+    () => getRulesPageData(filteredRulesWithStates, status, error, rulesQuery),
+    [filteredRulesWithStates, status, error, rulesQuery]
+  );
+
+  const setSelectAllRules = useCallback(() => {
+    setSelectedRules(rulesPageData.all_rules);
+  }, [rulesPageData.all_rules]);
+
+  const rulesFlyoutData: CspBenchmarkRulesWithStates = useMemo(() => {
+    const arrayRulesStates: RuleStateAttributes[] = Object.values(rulesStates.data || {});
+    return {
+      ...{
+        state:
+          arrayRulesStates.find((filteredRuleState) => filteredRuleState.rule_id === params.ruleId)
+            ?.muted === true
+            ? 'muted'
+            : 'unmuted',
+      },
+      ...{
+        metadata: allRules?.data?.items.find((rule) => rule.metadata.id === params.ruleId)
+          ?.metadata!,
+      },
+    };
+  }, [rulesStates, params.ruleId, allRules]);
+
+  const contextValue: RulesContextValue = useMemo(
     () => ({
-      allRules: allRules.data,
       rulesQuery,
+      rulesPageData,
       setRulesQuery: setRulesQueryCallback,
       pageSize,
       setPageSize: setPageSizeCallback,
       page,
       setPage: setPageCallback,
+      sectionList,
+      ruleNumberSelectOptions,
+      sectionSelectOptions,
+      selectedRules,
+      setSelectedRules,
+      setSelectAllRules,
+      setEnabledDisabledItemsFilter,
+      enabledDisabledItemsFilter,
+      mutedRulesCount,
+      rulesFlyoutData,
     }),
     [
-      allRules,
       rulesQuery,
+      rulesPageData,
       setRulesQueryCallback,
       pageSize,
       setPageSizeCallback,
       page,
       setPageCallback,
+      sectionList,
+      ruleNumberSelectOptions,
+      sectionSelectOptions,
+      selectedRules,
+      setSelectedRules,
+      setSelectAllRules,
+      setEnabledDisabledItemsFilter,
+      enabledDisabledItemsFilter,
+      mutedRulesCount,
+      rulesFlyoutData,
     ]
   );
 
-  return <RulesContext.Provider value={initialContextValue}>{children}</RulesContext.Provider>;
+  return <RulesContext.Provider value={contextValue}>{children}</RulesContext.Provider>;
 }
 
 export function useRules() {
